Mapper Update: Changed structure, Added venue and affiliation mapping functionalities
