--- src/fos_mapper/data/fos_labels_mapping_v0.1.2.json
@@ -1,34 +1,0 @@
-{
-    "settings": {
-        "number_of_shards": 8,
-        "number_of_replicas": 0
-    },
-    "mappings": {
-        "properties": {
-            "fos_label" : {
-                "type" : "text",
-                "fields" : {
-                    "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                    }
-                }
-            },
-            "level" : {
-                "type" : "text",
-                "fields" : {
-                  "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                  }
-                }
-            },
-            "fos_vector": {
-                "type": "dense_vector",
-                "dims": 768,
-                "index": "true",
-                "similarity": "cosine"
-            }
-        }
-    }
-}
