--- src/fos_mapper/data/fos_taxonomy_instruction_v0.1.2.json
@@ -1,4 +1,0 @@
-{
-    "embed_instruction": "search document: ",
-    "query_instruction": "search_query: "
-}
