--- src/fos_mapper/data/fos_taxonomy_mapping_v0.1.1.json
@@ -1,97 +1,0 @@
-{
-    "settings": {
-        "number_of_shards": 8,
-        "number_of_replicas": 0
-    },
-    "mappings": {
-        "properties": {
-            "level_1" : {
-                "type" : "text",
-                "fields" : {
-                    "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                    }
-                }
-            },
-            "level_2" : {
-                "type" : "text",
-                "fields" : {
-                  "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                  }
-                }
-            },
-            "level_3" : {
-                "type" : "text",
-                "fields" : {
-                  "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                  }
-                }
-            },
-            "level_4" : {
-                "type" : "text",
-                "fields" : {
-                  "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                  }
-                }
-            },
-            "level_4_id" : {
-                "type" : "text",
-                "fields" : {
-                  "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                  }
-                }
-            },
-            "level_5" : {
-                "type" : "text",
-                "fields" : {
-                  "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                  }
-                }
-            },
-            "level_5_name" : {
-                "type" : "text",
-                "fields" : {
-                  "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                  }
-                }
-            },
-            "level_5_id" : {
-                "type" : "text",
-                "fields" : {
-                  "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                  }
-                }
-            },
-            "level_6" : {
-                "type" : "text",
-                "fields" : {
-                  "keyword" : {
-                    "type" : "keyword",
-                    "ignore_above" : 256
-                  }
-                }
-            },
-            "fos_vector": {
-                "type": "dense_vector",
-                "dims": 768,
-                "index": "true",
-                "similarity": "cosine"
-            }
-        }
-    }
-}
